상품투자 API 개발

+++ src/main/java/com/kakaopay/ryuyungwang/investment/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.kakaopay.ryuyungwang.investment.repository;
 
-import com.kakaopay.ryuyungwang.investment.ProductEntity;
+import com.kakaopay.ryuyungwang.investment.entity.ProductEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
+public interface ProductRepository extends CrudRepository<ProductEntity, Integer> {
 
     List<ProductEntity> findAllByStartedAtLessThanEqualAndFinishedAtGreaterThanEqual(LocalDateTime now1, LocalDateTime now2);
 }
